phishingdetectionplatform/requirements.txt
absl-py==2.1.0
arrow==1.3.0
asgiref==3.8.1
astunparse==1.6.3
backports.zoneinfo==0.2.1
blessed==1.20.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
Django==4.2.17
django-picklefield==3.2
djangorestframework==3.15.2
flatbuffers==24.3.25
gast==0.4.0
google-auth==2.36.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.68.1
h5py==3.11.0
idna==3.10
importlib_metadata==8.5.0
keras==2.13.1
libclang==18.1.1
Markdown==3.7
MarkupSafe==2.1.5
numpy==1.24.3
oauthlib==3.2.2
opt_einsum==3.4.0
packaging==24.2
protobuf==4.25.5
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==3.5.3
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
sqlparse==0.5.2
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.5.0
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.6
wrapt==1.17.0
zipp==3.20.2
